requirements_streamlit_app.txt
streamlit>=1.36.0
pillow>=10.0.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3
werkzeug>=3.0.0
yagmail>=0.15.293
